% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{Bayes.1.Boot}
\alias{Bayes.1.Boot}
\title{Bayes.1.Boot}
\usage{
Bayes.1.Boot(xx, boot.iter, Seed)
}
\arguments{
\item{xx}{xx}

\item{boot.iter}{boot iteration}

\item{Seed}{seed}
}
\description{
Bayes.1.Boot
}
