% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{BB.MRL}
\alias{BB.MRL}
\title{BB.MRL}
\usage{
BB.MRL(S, M, AN, U, biter = 200, labnum = NULL, Seed, rnnr = 1)
}
\arguments{
\item{S}{S}

\item{M}{M}

\item{AN}{AB}

\item{U}{U}

\item{biter}{biter}

\item{labnum}{lab number}

\item{Seed}{seed}

\item{rnnr}{rnnr}
}
\description{
BB.MRL
}
