% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{compAllIRLS}
\alias{compAllIRLS}
\title{compAllIRLS}
\usage{
compAllIRLS(
  x,
  DX,
  DX2,
  DX3,
  DX4,
  y,
  VFnP,
  oldMSEfnPar1,
  oldB1,
  oldMSEfnPar2,
  oldB2,
  oldMSEfnPar3,
  oldB3,
  oldMSEfnPar4,
  oldB4,
  cc,
  k,
  RNNR,
  theta,
  maxIter,
  grphs,
  LN,
  BW
)
}
\arguments{
\item{x}{x}

\item{DX}{DX}

\item{DX2}{DX2}

\item{DX3}{DX3}

\item{DX4}{DX4}

\item{y}{y}

\item{VFnP}{VFnP}

\item{oldMSEfnPar1}{oldMSEfnPar1}

\item{oldB1}{oldB1}

\item{oldMSEfnPar2}{oldMSEfnPar2}

\item{oldB2}{oldB2}

\item{oldMSEfnPar3}{oldMSEfnPar3}

\item{oldB3}{oldB3}

\item{oldMSEfnPar4}{oldMSEfnPar4}

\item{oldB4}{oldB4}

\item{cc}{cc}

\item{k}{k}

\item{RNNR}{RNNR}

\item{theta}{theta}

\item{maxIter}{max iteration}

\item{grphs}{graphs}

\item{LN}{LN}

\item{BW}{BW}
}
\description{
compAllIRLS
}
