% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{compIRLS}
\alias{compIRLS}
\title{compIRLS}
\usage{
compIRLS(
  xx,
  DX,
  yy,
  VFnP,
  MSEFnP,
  B,
  cc,
  k,
  RNNR,
  theta,
  maxIter,
  grphs,
  LN,
  BW
)
}
\arguments{
\item{xx}{xx}

\item{DX}{DX}

\item{yy}{yy}

\item{VFnP}{VFnP}

\item{MSEFnP}{MSEFnP}

\item{B}{B}

\item{cc}{cc}

\item{k}{k}

\item{RNNR}{RNNR}

\item{theta}{theta}

\item{maxIter}{max iteration}

\item{grphs}{graph}

\item{LN}{LN}

\item{BW}{BW}
}
\description{
compIRLS
}
