% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{compMSE}
\alias{compMSE}
\title{compMSE}
\usage{
compMSE(X, Resids, cc, k, VFnP, BW)
}
\arguments{
\item{X}{x}

\item{Resids}{Resids}

\item{cc}{cc}

\item{k}{k}

\item{VFnP}{VFnP}

\item{BW}{BW}
}
\description{
compMSE
}
