% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{compWLS}
\alias{compWLS}
\title{compWLS}
\usage{
compWLS(x, DX, y, VFnPar, MSEFnP, oldB, cc, RNNR, BW)
}
\arguments{
\item{x}{x}

\item{DX}{DX}

\item{y}{y}

\item{VFnPar}{VFnPar}

\item{MSEFnP}{MSEFnP}

\item{oldB}{oldB}

\item{cc}{cc}

\item{RNNR}{RNNR}

\item{BW}{BW}
}
\description{
compWLS
}
