% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{DLEstFn}
\alias{DLEstFn}
\title{DLEstFn}
\usage{
DLEstFn(xd, yc, MFP, VFP, MSEFP, MV, b, RNNR)
}
\arguments{
\item{xd}{xd}

\item{yc}{yc}

\item{MFP}{MFP}

\item{VFP}{VFP}

\item{MSEFP}{MSEFP}

\item{MV}{MV}

\item{b}{b}

\item{RNNR}{RNNR}
}
\description{
DLEstFn
}
