% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{GammaEstFn}
\alias{GammaEstFn}
\title{GammaEstFn}
\usage{
GammaEstFn(xs, LQL, UQL, MFP, VFP, MSEFP, CPR, nn, xxbar, ssxx, RNNR)
}
\arguments{
\item{xs}{xs}

\item{LQL}{LQL}

\item{UQL}{UQL}

\item{MFP}{MFP}

\item{VFP}{VFP}

\item{MSEFP}{MSEFP}

\item{CPR}{CPR}

\item{nn}{nn}

\item{xxbar}{xxbar}

\item{ssxx}{ssxx}

\item{RNNR}{RNNR}
}
\description{
GammaEstFn
}
