% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{HubauxVos}
\alias{HubauxVos}
\title{HubauxVos}
\usage{
HubauxVos(MFP, VFP, MSEFP, a, b, lcmrlval, sL, RNNR, ZRF)
}
\arguments{
\item{MFP}{MFP}

\item{VFP}{VFP}

\item{MSEFP}{MSEFP}

\item{a}{a}

\item{b}{b}

\item{lcmrlval}{LCMRL value}

\item{sL}{sL}

\item{RNNR}{RNNR}

\item{ZRF}{ZRD}
}
\description{
HubauxVos
}
