% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{IRLS}
\alias{IRLS}
\title{IRLS}
\usage{
IRLS(xx, yy, VFnP, Wts, RNNR, theta, maxIter, grphs = 1, LN = NULL, BW)
}
\arguments{
\item{xx}{xx}

\item{yy}{yy}

\item{VFnP}{VFnP}

\item{Wts}{weights}

\item{RNNR}{RNNR}

\item{theta}{theta}

\item{maxIter}{max iteration}

\item{grphs}{graphs}

\item{LN}{NL}

\item{BW}{BW}
}
\description{
IRLS
}
