% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{LCMRL.data.rand.pert}
\alias{LCMRL.data.rand.pert}
\title{LCMRL.data.rand.pert}
\usage{
LCMRL.data.rand.pert(spike, measured, r.scale.sd = 0.03)
}
\arguments{
\item{spike}{spiked}

\item{measured}{measured}

\item{r.scale.sd}{r scale sd}
}
\description{
LCMRL.data.rand.pert
}
