% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{LCMRL.Values}
\alias{LCMRL.Values}
\title{LCRML.Values function}
\usage{
LCMRL.Values(
  fh,
  LQL = 0.5,
  UQL = 1.5,
  CPR = 0.99,
  alph = 0.05,
  bet = 0.05,
  rnnr = 1
)
}
\arguments{
\item{fh}{file name}

\item{LQL}{lower quantitation limit, default = 0.5}

\item{UQL}{upper quantitation limit, default = 1.5}

\item{CPR}{CPR}

\item{alph}{alpha error}

\item{bet}{beta error}

\item{rnnr}{set 1 if data consisting of only positive values, and 0 if data includes negative values}
}
\description{
LCRML.Values function
}
