% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrl4jecs.r
\docType{package}
\name{lcmrl4jecs}
\alias{lcmrl4jecs-package}
\alias{lcmrl4jecs}
\title{lcmrl4jecs}
\description{
This package is to impletement US EPA's LCMRL R code kindly provided by Dr Steven C. Wendelken.
}
\author{
\strong{Maintainer}: Shoji Nakayama \email{fabre@nies.go.jp}

}
