% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{MRL.Output}
\alias{MRL.Output}
\title{MRL.Output}
\usage{
MRL.Output(
  An,
  U,
  BB.tmp2,
  labflag,
  LCMRL.val,
  boot.iter,
  outfh,
  fh.a,
  Abbr = TRUE
)
}
\arguments{
\item{An}{An}

\item{U}{U}

\item{BB.tmp2}{BB.tmp2}

\item{labflag}{labflag}

\item{LCMRL.val}{LCMRL value}

\item{boot.iter}{number of iterations}

\item{outfh}{output file}

\item{fh.a}{file}

\item{Abbr}{abbreviations}
}
\description{
MRL.Output
}
