% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{MRL.Stats}
\alias{MRL.Stats}
\title{MRL.Stats}
\usage{
MRL.Stats(fh, rnnr = 1, boot.iter = 200, Seed = 261948, Abbr = TRUE)
}
\arguments{
\item{fh}{file}

\item{rnnr}{rnnr}

\item{boot.iter}{iteration number}

\item{Seed}{seed}

\item{Abbr}{abbreviation}
}
\description{
MRL.Stats
}
