% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{MRL.Summary}
\alias{MRL.Summary}
\title{MRL LCMRL estimatation scripts}
\usage{
MRL.Summary(fh.labdata, rnnr = 1, biter = 200, Seed = 261948, abbr.out = TRUE)
}
\arguments{
\item{fh.labdata}{lab data}

\item{rnnr}{rnnr}

\item{biter}{biter}

\item{Seed}{seed}

\item{abbr.out}{abbreviation}
}
\description{
MRL.Summary
}
