% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{OptABC}
\alias{OptABC}
\title{OptABC}
\usage{
OptABC(A, B, CC, SL, RV, RVDF)
}
\arguments{
\item{A}{A}

\item{B}{B}

\item{CC}{CC}

\item{SL}{SL}

\item{RV}{RV}

\item{RVDF}{RVDF}
}
\description{
OptABC
}
