% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfas}
\alias{pfas}
\title{PFAS data}
\format{
An object of class \code{data.frame} with 544 rows and 6 columns.
}
\usage{
pfas
}
\description{
A dataset containing PFAS data
}
\keyword{datasets}
