% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{robCondMSE}
\alias{robCondMSE}
\title{robCondMSE}
\usage{
robCondMSE(X, Resids, cc, k, BW)
}
\arguments{
\item{X}{x}

\item{Resids}{Resids}

\item{cc}{cc}

\item{k}{k}

\item{BW}{BW}
}
\description{
robCondMSE
}
