% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{RobLCMRL}
\alias{RobLCMRL}
\title{RobLCMRL}
\usage{
RobLCMRL(
  SpikeConc,
  MeasConc,
  AnalName,
  strUnits = "ug/L",
  lowQL = 0.5,
  upQL = 1.5,
  CovProbReq = 0.99,
  alpha = 0.05,
  beta = 0.05,
  ReqNonNegResponse = 1,
  grphs = 1,
  labnum = NULL,
  BW = rep(1, length(SpikeConc))
)
}
\arguments{
\item{SpikeConc}{spiked concentration}

\item{MeasConc}{measured concentration}

\item{AnalName}{analyte names}

\item{strUnits}{Units}

\item{lowQL}{lower qunatiation limit}

\item{upQL}{upper quantitation limit}

\item{CovProbReq}{covprobreq}

\item{alpha}{alpha}

\item{beta}{beta}

\item{ReqNonNegResponse}{required non negative response}

\item{grphs}{graphs}

\item{labnum}{lab number}

\item{BW}{BW}
}
\description{
RobLCMRL
}
