% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{roblocvar4}
\alias{roblocvar4}
\title{roblocvar4}
\usage{
roblocvar4(XX, cc = 9, k = 1, convtype = 0, Pwts = rep(1, length(XX)))
}
\arguments{
\item{XX}{XX}

\item{cc}{cc}

\item{k}{k}

\item{convtype}{conversion type}

\item{Pwts}{Pwts}
}
\description{
roblocvar4
}
