% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{RobRepVar}
\alias{RobRepVar}
\title{RobRepVar}
\usage{
RobRepVar(X, Y, cc = 9, k = 1, aPwts)
}
\arguments{
\item{X}{X}

\item{Y}{Y}

\item{cc}{cc}

\item{k}{k}

\item{aPwts}{aPwts}
}
\description{
RobRepVar
}
