% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{srchDL}
\alias{srchDL}
\title{srchDL}
\usage{
srchDL(lcmrlval, yc, sL, MFP, VFP, MSEFP, MV, b, RNNR, ZRF)
}
\arguments{
\item{lcmrlval}{lcmrl value}

\item{yc}{yc}

\item{sL}{sL}

\item{MFP}{MFP}

\item{VFP}{VFP}

\item{MSEFP}{MSEFP}

\item{MV}{MV}

\item{b}{beta}

\item{RNNR}{RNNR}

\item{ZRF}{ZRF}
}
\description{
srchDL
}
