% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{srchLCMRL}
\alias{srchLCMRL}
\title{srchLCMRL}
\usage{
srchLCMRL(SL, LQL, UQL, MFP, VFP, MSEFP, CPR, nn, xxbar, ssxx, RNNR, LLLCMRL)
}
\arguments{
\item{SL}{SL}

\item{LQL}{LQL}

\item{UQL}{UQL}

\item{MFP}{MFP}

\item{VFP}{VFP}

\item{MSEFP}{MSEFP}

\item{CPR}{CPR}

\item{nn}{nn}

\item{xxbar}{xxbar}

\item{ssxx}{ssxx}

\item{RNNR}{RNNR}

\item{LLLCMRL}{LLLCMRL}
}
\description{
srchLCMRL
}
