% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test}
\alias{test}
\title{Test data}
\format{
An object of class \code{data.frame} with 220 rows and 6 columns.
}
\usage{
test
}
\description{
A dataset containing test data
}
\keyword{datasets}
