% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{wgt.guttman.utl}
\alias{wgt.guttman.utl}
\title{wgt.guttman.utl}
\usage{
wgt.guttman.utl(x, y, a = 0.95, b = 0.95)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{a}{alpha}

\item{b}{beta}
}
\description{
wgt.guttman.utl
}
