% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmrlscripts.r
\name{wt.bw.1.sided}
\alias{wt.bw.1.sided}
\title{wt.bw.1.sided}
\usage{
wt.bw.1.sided(x, rloc, rscale, cc = 6, reltol = 1e-06, maxsteps = 100)
}
\arguments{
\item{x}{x}

\item{rloc}{rloc}

\item{rscale}{rscale}

\item{cc}{cc}

\item{reltol}{relocation tolerance}

\item{maxsteps}{max steps}
}
\description{
wt.bw.1.sided
}
